v4.0.1.0 代码迁移到vs2010 优化部分bug处理

+++ SerialDebug/FormFileSend.cs
@@ -65,7 +65,7 @@ namespace SerialDebug
             Properties.Settings.Default.filePath = txtFile.Text;
             Properties.Settings.Default.fileSendByLine = chkSendByLine.Checked;
             Properties.Settings.Default.fileSendCRLF = chkSendCRRF.Checked;
-            (int)Properties.Settings.Default.fileSendDelay = (int)numDelayTime.Value;
+            Properties.Settings.Default.fileSendDelay = (int)numDelayTime.Value;
             Properties.Settings.Default.fileSendByPacket = chkSlipPacket.Checked;
             Properties.Settings.Default.filePacketLen = (int)numPacketLen.Value;
